+++ ppa/package.sh
@@ -2,6 +2,12 @@
 
 SCRIPTPATH=$(dirname "$0") 
 VERSION=$(cat ${SCRIPTPATH}/../VERSION)
+
+if [ -f  ${SCRIPTPATH}/../keyid ]; then
+    echo "Export Keyid from file"
+    export DEB_SIGN_KEYID=$(cat ${SCRIPTPATH}/../keyid)
+fi
+
 DISTROS=("jammy" "noble")
 
 SRC_FOLDER=mkdotenv_${VERSION}
@@ -16,13 +22,14 @@ DIST=jammy
 bash ${SCRIPTPATH}/make_tar.sh
 
 cd ${SCRIPTPATH}/..
-
+pwd
+sleep 10
 for distro in "${DISTROS[@]}"; do
     echo "Create source package for: "${distro}
 
     sed -i "s/unstable/${distro}/g" ${SCRIPTPATH}/../debian/changelog
 	sed -i 's/debian/ubuntu/g' ${SCRIPTPATH}/../debian/changelog
 	dpkg-buildpackage -S -sa
-	sed -i "s/${distro}/unstable/g" ${SCRIPTPATH}/debian/changelog
-	sed -i 's/ubuntu/debian/g' ${SCRIPTPATH}/debian/changelog
+	sed -i "s/${distro}/unstable/g" ${SCRIPTPATH}/../debian/changelog
+	sed -i 's/ubuntu/debian/g' ${SCRIPTPATH}/../debian/changelog
 done