+++ ppa/testbuild.sh
@@ -2,9 +2,8 @@
 
 set -e  # Exit on any error
 
-SCRIPTPATH=$(dirname "$0")
+SCRIPTPATH=$(dirname "$0") 
 bash ${SCRIPTPATH}/package.sh
-
 VERSION=$(cat ${SCRIPTPATH}/../VERSION)
 
 ls -l ${SCRIPTPATH}/../../mkdotenv_${VERSION}.orig.tar.gz
@@ -18,5 +17,13 @@ echo "Testbuild source packages"
 DISTROS=("jammy" "noble")
 MIRROR="http://archive.ubuntu.com/ubuntu"
 
+PPA_REPO="ppa:pcmagas/mkdotenv"
+
+if [ -f ${SCRIPTPATH}/../PPA_OVERRIDE ]; then
+    PPA_REPO=$(cat ${SCRIPTPATH}/../PPA_OVERRIDE)
+fi
+
+echo "UPLOAD into ${PPA_REPO}"
+sleep 5
 rm -rf ${SCRIPTPATH}/../../mkdotenv_*.ppa.upload 
-dput ppa:pcmagas/ppa-test5 ${SCRIPTPATH}/../../mkdotenv_${VERSION}-0ubuntu1~*1_source.changes
+dput ${PPA_REPO} ${SCRIPTPATH}/../../mkdotenv_${VERSION}-0ubuntu1~*1_source.changes